Replace defaultValue props of option tags into value and prevent future bugs
When there's no value prop, formik would track the value of the gender select field by looking at the contents of the option tags which can
vary from language to language, we didn't wanna translate the database also, we just wanted to translate the UI so we should have use the
English version of the data and show that data in the version that user wanted by looking at the language that they have selected. But in
this case it was writing the Turkish versions of the gender field to the server which is not what we wanted.